Change plot value precision

+++ mmperf.py
@@ -114,7 +114,7 @@ def autolabel(rects):
     for rect in rects:
         height = rect.get_height()
         plt.text(rect.get_x() + rect.get_width()/2., 1.02*height,
-                '%d' % int(height), fontsize=5, ha='center', va='bottom')
+                '%.3f' % float(height), fontsize=5, ha='center', va='bottom')
 
 _result_dir = None
 _env = None
